kisankiawaz-backend/services/crop/requirements.txt
# Crop-specific deps
Pillow>=10.0.0
google-generativeai>=0.8.0
